Split up filters.py into one file per filter class
